factors out repeated code for unescaping commas #154
