+++ lib/facets/aggs-objects.js
@@ -1,3 +1,5 @@
+const unescapeCommas = require('../unescape-commas');
+
 module.exports = function (queryParams) {
   const aggregationObjects = {
     filter: {
@@ -112,15 +114,5 @@ function filter (queryParams, exclude) {
     }
   }
 
-  return filters.map(el => {
-    var filt = {
-      terms: {}
-    };
-    if (el.terms && Array.isArray(el.terms[Object.keys(el.terms)[0]])) {
-      filt.terms[Object.keys(el.terms)[0]] = el.terms[Object.keys(el.terms)[0]].map(el => {
-        return el.replace(/\\,/g, ',');
-      });
-    }
-    return filt;
-  });
+  return filters.map(unescapeCommas);
 }
