+++ lib/facets/aggs-people.js
@@ -1,3 +1,4 @@
+const unescapeCommas = require('../unescape-commas');
 /**
 * Create an aggregation object for creating facets
 */
@@ -74,15 +75,5 @@ function filter (queryParams, exclude) {
     }
   }
 
-  return filters.map(el => {
-    var filt = {
-      terms: {}
-    };
-    if (el.terms && Array.isArray(el.terms[Object.keys(el.terms)[0]])) {
-      filt.terms[Object.keys(el.terms)[0]] = el.terms[Object.keys(el.terms)[0]].map(el => {
-        return el.replace(/\\,/g, ',');
-      });
-    }
-    return filt;
-  });
+  return filters.map(unescapeCommas);
 }
