+++ lib/facets/create-filter-all.js
@@ -1,3 +1,4 @@
+const unescapeCommas = require('../unescape-commas');
 /**
 * Create an array of terms to filter objects in elasticsearch
 */
@@ -47,15 +48,5 @@ module.exports = function (queryParams) {
     filters.push({ terms: { 'agents.summary_title': user } });
   }
 
-  return filters.map(el => {
-    var filt = {
-      terms: {}
-    };
-    if (el.terms && Array.isArray(el.terms[Object.keys(el.terms)[0]])) {
-      filt.terms[Object.keys(el.terms)[0]] = el.terms[Object.keys(el.terms)[0]].map(el => {
-        return el.replace(/\\,/g, ',');
-      });
-    }
-    return filt;
-  });
+  return filters.map(unescapeCommas);
 };
