+++ lib/facets/create-filter-documents.js
@@ -1,3 +1,5 @@
+const unescapeCommas = require('../unescape-commas');
+
 module.exports = function (queryParams) {
   var filters = [];
 
@@ -12,15 +14,5 @@ module.exports = function (queryParams) {
     filters.push({ terms: {'organisations.name.value': organisations} });
   }
 
-  return filters.map(el => {
-    var filt = {
-      terms: {}
-    };
-    if (el.terms && Array.isArray(el.terms[Object.keys(el.terms)[0]])) {
-      filt.terms[Object.keys(el.terms)[0]] = el.terms[Object.keys(el.terms)[0]].map(el => {
-        return el.replace(/\\,/g, ',');
-      });
-    }
-    return filt;
-  });
+  return filters.map(unescapeCommas);
 };
