+++ lib/facets/create-filter-people.js
@@ -1,3 +1,4 @@
+const unescapeCommas = require('../unescape-commas');
 /**
 * Create an array of terms for filter with elasticsearch
 * [{terms: {'occupation': [val1, val2,...]}}, {...}]
@@ -34,15 +35,5 @@ module.exports = function (queryParams) {
     }
   }
 
-  return filters.map(el => {
-    var filt = {
-      terms: {}
-    };
-    if (el.terms && Array.isArray(el.terms[Object.keys(el.terms)[0]])) {
-      filt.terms[Object.keys(el.terms)[0]] = el.terms[Object.keys(el.terms)[0]].map(el => {
-        return el.replace(/\\,/g, ',');
-      });
-    }
-    return filt;
-  });
+  return filters.map(unescapeCommas);
 };
